pyconsumer/requirements.txt
kafka-python==2.0.2
mysql-connector-python==8.0.27
protobuf==3.19.1
PyMySQL==1.0.2
python-dateutil==2.8.2
six==1.16.0
text-unidecode==1.3
typing-extensions==4.0.0
